+++ Samples~/BasicUsage/MovementTrackerSample.cs
@@ -6,9 +6,21 @@ namespace Jeomseon.Samples.Components
     [RequireComponent(typeof(MovementTracker))]
     public sealed class MovementTrackerSample : MonoBehaviour
     {
+        private MovementTracker _movementTracker;
+
+        private void Awake()
+        {
+            _movementTracker = GetComponent<MovementTracker>();
+        }
+
         private void OnEnable()
         {
-            GetComponent<MovementTracker>().AddListenerOnMoveBegined(OnMoveBegan);
+            _movementTracker.MoveBegan += OnMoveBegan;
+        }
+
+        private void OnDisable()
+        {
+            _movementTracker.MoveBegan -= OnMoveBegan;
         }
 
         private static void OnMoveBegan(Vector3 delta)